extract caching mechanism from CheckCandidates
The code of phase 1 maintains a cache `regex_map` to avoid checking
several times the same regex. But for some reason it was checked inside
the check loop rather than before starting it, which is pretty confusing
and obscure the algorithm.

This fixes the order and make sure the cache is always updated at the
same location in the code.